# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VelozTrade API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List trade history
      description: Returns closed/historical trades ordered by close time desc
      parameters:
        - name: symbol
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Trade history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}/journal:
    patch:
      operationId: updateOrderJournal
      tags: [orders]
      summary: Update trade journal entry
      description: Attach or update notes, strategy tag, and sentiment rating on a closed trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /pending-orders:
    get:
      operationId: listPendingOrders
      tags: [pendingOrders]
      summary: List pending limit orders
      description: Returns all pending limit orders for the authenticated user
      responses:
        "200":
          description: Pending orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingOrder"
    post:
      operationId: createPendingOrder
      tags: [pendingOrders]
      summary: Place a pending limit order
      description: Creates a limit order that activates when the market price reaches the limit price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PendingOrderInput"
      responses:
        "201":
          description: Pending order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingOrder"
        "400":
          description: Invalid order parameters
  /pending-orders/{id}:
    delete:
      operationId: cancelPendingOrder
      tags: [pendingOrders]
      summary: Cancel a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled
        "404":
          description: Pending order not found
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List price alerts
      description: Returns all active and triggered alerts for the current user
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceAlert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a price alert
      description: Set a price alert that fires when a symbol crosses the target price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceAlertInput"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
        "400":
          description: Invalid input
  /alerts/{id}:
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete a price alert
      description: Remove an alert by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Alert not found
...
  /alerts/{id}/trigger:
    patch:
      operationId: triggerAlert
      tags: [alerts]
      summary: Mark alert as triggered
      description: Called by the client when the WS price crosses the threshold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceAlert"
        "404":
          description: Alert not found
components:
  schemas:
...
    Order:
      type: object
      required: [id, symbol, direction, volume, openPrice, closePrice, profit, openTime, closeTime]
      properties:
        id:
          type: integer
        symbol:
          type: string
        symbolName:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        volume:
          type: number
        openPrice:
          type: number
        closePrice:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        profit:
          type: number
        swap:
          type: number
        commission:
          type: number
        openTime:
          type: string
        closeTime:
          type: string
        journalNote:
          type: ["string", "null"]
        strategyTag:
          type: ["string", "null"]
        sentimentRating:
          type: ["integer", "null"]
    JournalInput:
      type: object
      properties:
        journalNote:
          type: ["string", "null"]
        strategyTag:
          type: ["string", "null"]
        sentimentRating:
          type: ["integer", "null"]
    PendingOrder:
      type: object
      required: [id, symbol, direction, orderType, volume, limitPrice, marginReserved, status, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        symbolName:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        orderType:
          type: string
          enum: [buy_limit, sell_limit, buy_stop, sell_stop, buy_stop_limit, sell_stop_limit, trailing_stop]
        volume:
          type: number
        limitPrice:
          type: number
        stopPrice:
          type: ["number", "null"]
        trailingDistance:
          type: ["number", "null"]
        marginReserved:
          type: number
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, filled, cancelled]
        createdAt:
          type: string
    PendingOrderInput:
      type: object
      required: [symbol, direction, volume, limitPrice]
      properties:
        symbol:
          type: string
        direction:
          type: string
          enum: [buy, sell]
        orderType:
          type: string
          enum: [buy_limit, sell_limit, buy_stop, sell_stop, buy_stop_limit, sell_stop_limit, trailing_stop]
        volume:
          type: number
        limitPrice:
          type: number
        stopPrice:
          type: ["number", "null"]
        trailingDistance:
          type: ["number", "null"]
        stopLoss:
          type: ["number", "null"]
        takeProfit:
          type: ["number", "null"]
...
    PriceAlert:
      type: object
      required: [id, symbol, targetPrice, condition, status, createdAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        targetPrice:
          type: number
        condition:
          type: string
          enum: [above, below]
        status:
          type: string
          enum: [active, triggered, cancelled]
        note:
          type: ["string", "null"]
        createdAt:
          type: string
        triggeredAt:
          type: ["string", "null"]
    PriceAlertInput:
      type: object
      required: [symbol, targetPrice, condition]
      properties:
        symbol:
          type: string
        targetPrice:
          type: number
        condition:
          type: string
          enum: [above, below]
        note:
          type: ["string", "null"]
...